merge: return number of acceptable configs
We warn but don't fail on invalid messages; returning the number of
accepted ones can let the client decide if they want to treat it more
seriously or not.

(Clients probably shouldn't hard fail: someone else, e.g. a malicious
actor or service node, could have injected a bad message, in which case
the client should just ignore it).