Fix dirty->merge->dirty mutability bug
If you have a dirty config and merged pushed changes (before pushing
yourself) you'd end up with a non-mutable internal config object that is
marked dirty.  It *should* be dirty (because the dirtyness carries
forward from before the merge), but it should also be Mutable because
dirty config objects are always expected/required to be mutable.

This fixes it by properly constructing a MutableConfigMessage instead of
just a ConfigMessage when we merge configs into an already-dirty config
so that we properly remain mutable.