Fix dirty-but-immutable state bug
If we merge configs that triggers a merge message, we would get in a
state where we *don't* have a mutable config message, but still need
things to be marked dirty (because of the merge), which is an invalid
state.

This fixes it: if such a case occurs we now convert the ConfigMessage
into a MutableConfigMessage (*without* incrementing its seqno, since
that is already done as part of the merge), which leaves everything in a
proper dirty state *and* lets us make more changes (if needed) to push
the merge + new changes all at once.

Includes a test case that reproduces it when run without the fixes made here.