Fix unnecessary seqno increment + assert failure
If we are currently in a dirty state and merge a config that ends up
being useless to us (i.e. we already have it, or its invalid) then we
will load *only* from the current config, but will load into a
MutableConfigMessage which always increments seqno.

We should ignore that case and drop the "new" config, though, because
it's no different than our current config (it just wastes a seqno).  The
assert just after this change was triggering, *correctly*, because we
aren't supposed to be doing seqno increments when we don't change
anything through the merge.

This fixes the increment *and* avoids the assert by detecting this case
and doing nothing when it happens.