Fix assigning empty dicts/sets dirtying
When assigning an empty set/dict to a location that doesn't currently
exist, we'd dirty the config, assign the empty dict/set, but then when
we produce the update it is unchanged because empty dicts/sets get
pruned.

This fixes it so that such assignment becomes an *erase* instead, since
that what the final result will be.