Update android ndk deb
Getting this error from the CI
```
Package google-android-ndk-installer is not available, but is referred to by another package.

This may mean that the package is missing, has been obsoleted, or

is only available from another source

However the following packages replace it:

  google-android-ndk-r25c-installer
```

android-ndk-r25c is the current long term release according to
https://github.com/android/ndk/wiki and it appears that the source
repos have depreciated the general package for specific versions. This
has been included in the docker images already so we dont need to
include the ndk here anymore