Experimental zstd training
This attempt (trained on various real-world config data) only noticeably
improved compression of tiny files (e.g. on the order of hundreds of
bytes), and had virtually no improvements on larger files (which is
really where we care).

This commit includes it for posterity, but it won't get included in
libsession-util.