Remove broken erase(iterator) interface
Erasure can cause config dirtying, which invalidates not only all
current iterators, but also the current dict object and isn't something
we can realistically manage while iterating.

Currently, if this happened, we would crash, which is bad.

This removes the iterator-erase interface completely to avoid the issue,
and updates documentation to state that modifications while iterating
are *NOT* permitted.