Group metadata; signed & read-only configs support
Adds a group info config type, which tracks distributed info for v2
groups.

This requires introducing/using a few new concepts not currently used
for user config messages:

- Multiple decryption keys.  User config doesn't do this at all (rather
  it generates a single decryption key from the private key for each
  namespace).

This doesn't yet add support for sharing and distributing those keys,
just for having being able to load a config with a list of multiple
possible keys.

- Config signing and verification.  For user configs this isn't
  done, since only the owner can actually encrypt/decrypt a config
  message, just being able to decrypt it is authentication enough.

This required various modifications to make the config library properly
prevent modifications when we can't modify, and to properly follow
that through in terms of merging, updates, etc.