Remove `signature_optional` parameter
It doesn't really make sense to have this as we either want signatures
(for shared messages) or don't (for personally encrypted messages).  The
only place we were passing it as `true` was in a place that also didn't
pass a verifier or signer (during dumping), and so already wasn't
adding/checking signatures.