Add invited and name to groups; add kicked methods
This adds an `.invited` flag for all group types (legacy, new, and
communities) that can be used to track a invited-but-not-yet-joined room
by session clients.

Also adds `.name` to new groups data so that the name from an invitation
can be stored before accepting (after accepting the name will come from
the shared group config message).

Add `setKicked()` and `kicked()` methods which clear both auth_data and
secretkey and report whether both are empty, respectively.  (For the C
API these are `ugroups_group_set_kicked` and `ugroups_group_is_kicked`
and take a pointer to the `ugroups_group_info`).