Add onion req parsing/encryption support
This adds support for onion request encryption/decryption and onion
request final request parsing and replying.

This code is moving over from libonionrequest, though with some
significant rewriting:
- the request parser object structure is significantly slimmed down
- the AES256-GCM implementation reimplemented in nettle rather than
  depending on OpenSSL for it
- AES256-CBC is dropped entirely; we technically still support it but
  have never actually used it, because it sucks.  This implementation
  supports just the AES256-GCM and XChaCha20-Poly1305 encryptions.

The nettle dependency is also required by libquic for gnutls, which will
likely eventually become a dependency of this code, so this move away
from OpenSSL aligns with that change.