Add deterministic session-encryption; use it for config protobuf
This adds a deterministic version of session protocol encryption that
differs from crypto_box_seal by using a keyed hash of the sender seed,
recipient pubkey, and message to generate the ephemeral key.

This effectively neuters such an encrypted message to be a regular
crypto_box encryption, which means the sender is later able to decrypt
it; this is thus only really a special case for protobuf+encrypted
config messages for backwards compatible config message support.