Fix Android build logic
Modern cmake + modern android versions let us be a bit less hacky with
getting things for the Android build system, so do that.  This should
also fix compilation for Android on macos.

This also fixes it to properly use ccache for android builds (previously
the ccache prefix on the compiler was set too early and being
obliterated under Android).